+++ railties/test/application/configuration_test.rb
@@ -1497,8 +1497,10 @@ module ApplicationTests
       end
     end
 
-    test "gc is disabled during autoload init" do
+    test "gc is disabled when config.disable_gc is true" do
       add_to_config <<-RUBY
+        config.disable_gc = true
+
         initializer :check_gc_enabled, after: :disable_gc do
           ::GC_WAS_DISABLED = GC.disable
         end
@@ -1507,20 +1509,6 @@ module ApplicationTests
       app "development"
 
       assert GC_WAS_DISABLED
-    end
-
-    test "gc is enabled during eager load init" do
-      add_to_config <<-RUBY
-        config.eager_load = true
-
-        initializer :check_gc_enabled, after: :disable_gc do
-          ::GC_WAS_ENABLED = !GC.enable
-        end
-      RUBY
-
-      app "development"
-
-      assert GC_WAS_ENABLED
     end
 
     test "valid beginning of week is setup correctly" do